+++ tests/entrypoints/openai/responses/test_basic.py
