+++ tests/entrypoints/openai/responses/test_function_call.py
