+++ tests/entrypoints/openai/responses/test_image.py
