+++ tests/entrypoints/openai/responses/test_structured_output.py
